+++ Content.Shared/_FarHorizons/Silicons/IPC/Traits/Negative/NegativeCustomFrameTraitSystem.cs
@@ -1,35 +1,17 @@
 using Content.Shared._FarHorizons.IPC.Traits;
 using Content.Shared._FarHorizons.Silicons.IPC.Components;
 using Content.Shared.Containers.ItemSlots;
+using Content.Shared.Movement.Systems;
 using Content.Shared.Tag;
 using Content.Shared.Whitelist;
 using Robust.Shared.Prototypes;
 
 namespace Content.Shared._FarHorizons.Silicons.IPC.Traits.Negative;
 
-public sealed class BloodPoweredTraitTraitSystem : IPCTraitSystem<BloodPoweredTraitComponent>
+public sealed class MicroreactorIncompatibilityTraitSystem : IPCTraitSystem<MicroreactorIncompatibilityTraitComponent>
 {
     [Dependency] private readonly ItemSlotsSystem _itemSlots = default!;
-    protected override void TraitInit(Entity<IPCBrainHolderComponent, BloodPoweredTraitComponent> ent)
-    {
-        if(!TryComp<IPCBatteryComponent>(ent.Owner, out var ipcBattery))
-            return;
-
-        ipcBattery.DrainAllowedTargets.Clear();
-        Dirty(ent.Owner, ipcBattery);
-        
-        if(!_itemSlots.TryGetSlot(ent.Owner, ipcBattery.PowerCellSlot.CellSlotId, out var cellSlot))
-            return;
-
-        _itemSlots.SetDisableEject(ent.Owner, cellSlot, true);
-        _itemSlots.SetSwap(ent.Owner, cellSlot, false);
-    }
-}
-
-public sealed class MicroreactorIncompatibilityTraitSystem : IPCTraitSystem<MicroreactorIncompatibilityComponent>
-{
-    [Dependency] private readonly ItemSlotsSystem _itemSlots = default!;
-    protected override void TraitInit(Entity<IPCBrainHolderComponent, MicroreactorIncompatibilityComponent> ent)
+    protected override void TraitInit(Entity<IPCBrainHolderComponent, MicroreactorIncompatibilityTraitComponent> ent)
     {
         if(!TryComp<IPCBatteryComponent>(ent.Owner, out var ipcBattery) || 
             !_itemSlots.TryGetSlot(ent.Owner, ipcBattery.PowerCellSlot.CellSlotId, out var cellSlot))
@@ -43,4 +25,19 @@ public sealed class MicroreactorIncompatibilityTraitSystem : IPCTraitSystem<Micr
         blackList.Tags.Add("PowerCellMicroreactor");
         _itemSlots.SetBlacklist(ent.Owner, cellSlot, blackList);
     }
+}
+
+public sealed class HeavierFrameTraitSystem: IPCTraitSystem<HeavierFrameTraitComponent>
+{
+    [Dependency] private readonly MovementSpeedModifierSystem _speedModifier = default!;
+    public override void Initialize()
+    {
+        base.Initialize();
+        SubscribeLocalEvent<HeavierFrameTraitComponent, RefreshMovementSpeedModifiersEvent>(OnRefreshMovementSpeed);
+    }
+    protected override void TraitInit(Entity<IPCBrainHolderComponent, HeavierFrameTraitComponent> ent) 
+        => _speedModifier.RefreshMovementSpeedModifiers(ent);
+
+    private void OnRefreshMovementSpeed(Entity<HeavierFrameTraitComponent> ent, ref RefreshMovementSpeedModifiersEvent args) 
+        => args.ModifySpeed(ent.Comp.SpeedModifier);
 }
