+++ Content.Shared/_FarHorizons/Silicons/IPC/Traits/Positive/PositiveCustomFrameTraitSystem.cs
@@ -107,3 +107,22 @@ public sealed class RepairNanitesTraitSystem : IPCToggleActionTraitSystem<Repair
         }
     }
 }
+
+public sealed class BloodPoweredTraitTraitSystem : IPCTraitSystem<BloodPoweredTraitComponent>
+{
+    [Dependency] private readonly ItemSlotsSystem _itemSlots = default!;
+    protected override void TraitInit(Entity<IPCBrainHolderComponent, BloodPoweredTraitComponent> ent)
+    {
+        if(!TryComp<IPCBatteryComponent>(ent.Owner, out var ipcBattery))
+            return;
+
+        ipcBattery.DrainAllowedTargets.Clear();
+        Dirty(ent.Owner, ipcBattery);
+        
+        if(!_itemSlots.TryGetSlot(ent.Owner, ipcBattery.PowerCellSlot.CellSlotId, out var cellSlot))
+            return;
+
+        _itemSlots.SetDisableEject(ent.Owner, cellSlot, true);
+        _itemSlots.SetSwap(ent.Owner, cellSlot, false);
+    }
+}
