+++ Resources/Locale/en-US/_FarHorizons/traits/custom_frame-traits.ftl
@@ -2,9 +2,6 @@
 custom-frame-trait-overvolted-i-name = Overvolted I
 custom-frame-trait-overvolted-i-description = Your frame was built with a slightly overvolted power supply. (+25% power drain)
 
-custom-frame-trait-overvolted-ii-name = Overvolted II
-custom-frame-trait-overvolted-ii-description = Your frame was built with a overvolted power supply. (+50% power drain)
-
 custom-frame-trait-cheap-oil-filter-name = Cheap Oil Filter
 custom-frame-trait-cheap-oil-filter-description = Your frame was built with a cheap oil filter making it unable to recycle oil. (Disables oil regeneration)
 
@@ -14,12 +11,11 @@ custom-frame-trait-compromised-integrity-description = Your frame was built with
 custom-frame-trait-radiation-vulnerability-name = Radiation Vulnerability
 custom-frame-trait-radiation-vulnerability-description = Your frame was built with a cheaper grade of radiation proctection. (-30% Radiation resistance and +50% duration to Ion storm glitches)
 
-custom-frame-trait-blood-powered-name = Blood Powered
-custom-frame-trait-blood-powered-description = Your frame was build with a rather esoteric battery its capable of minor repairs and recharging on its own, but at the cost of requiring blood. (Blood heals you and recharges your batter also batteries aren't swappable and you lose the ability to recharge from APCs)
-custom-frame-trait-blood-powered-healed = You feel the blood restoring your body.
-
 custom-frame-trait-microreactor-incompatibility-name = Microreactor Incompatibility
 custom-frame-trait-microreactor-incompatibility-description = Your frame was built in delicate internals that cannot withstand the radiation from a microreactor. (Disables microreactors insertion)
+
+custom-frame-trait-heavier-frame-name = Heavier Frame
+custom-frame-trait-heavier-frame-description = Your frame was built with a cheaper but heavier plasteel alloy to compensate your servos are stronger but louder but even with the stronger servos you are slightly slower. (Louder footsteps and 5% slowdown)
 
 #Positives
 custom-frame-trait-undervolted-i-name = Undervolted I
@@ -31,9 +27,6 @@ custom-frame-trait-undervolted-ii-description = Your frame was built with a unde
 custom-frame-trait-repair-nanites-name = Repair Nanites
 custom-frame-trait-repair-nanites-description = Your frame was built with an in-built nanite replicator capable of doing repairs on command but at the cost of extra power. (Adds ability that can toggle regeneration on and off)
 
-custom-frame-trait-deluxe-oil-filter-name = Deluxe Oil Filter
-custom-frame-trait-deluxe-oil-filter-description = Your frame was built with a deluxe oil filter its capable of clotting wounds and faster oil recycling. (Enables blood clotting and increased oil regeneration)
-
 custom-frame-trait-module-compatibility-name = Module Compatibility
 custom-frame-trait-module-compatibility-description = Your frame was built with the ability to support cyborg modules. (Allows cyborg modules to be installed)
 
@@ -41,4 +34,8 @@ custom-frame-trait-overclocking-name = Overclocking
 custom-frame-trait-overclocking-description = Your frame was built with the ability to redirect power from your fans directly to your servos. (Adds ability that shuts down your cooling fans but increases your speed by x1.35 and cosnumes some power)
 
 custom-frame-trait-better-insulation-name = Better Insulation
-custom-frame-trait-better-insulation-description = Your frame was built with better insulating materials making you harder to be damaged by the cold and heat as well less effects from shock. (+20% Heat, +20% Cold and +20% Shock. And minor resistance to electrocution.)
+custom-frame-trait-better-insulation-description = Your frame was built with better insulating materials making you harder to electrocute. (Adds electrocution protection. Not full insulated gloves!)
+
+custom-frame-trait-blood-powered-name = Blood Powered
+custom-frame-trait-blood-powered-description = Your frame was build with a rather esoteric battery its capable of minor repairs and recharging on its own, but at the cost of requiring blood. (Blood heals you and recharges your battery also batteries aren't swappable and you lose the ability to recharge from APCs)
+custom-frame-trait-blood-powered-healed = You feel the blood restoring your body.
