+++ Resources/Prototypes/_FarHorizons/Damage/modifier_sets.yml
@@ -150,16 +150,6 @@
     Piercing: 0.7
 
 - type: damageModifierSet
-  id: IPCInsulated
-  coefficients:
-    Shock: 1.3
-    Blunt: 1.2 
-    Slash: 0.7
-    Piercing: 0.7
-    Heat: 0.8
-    Cold: 0.8
-
-- type: damageModifierSet
   id: IPCWeaker
   coefficients:
     Shock: 1.5
