+++ Resources/Prototypes/_FarHorizons/Traits/Custom_Frame/custom_frame_traits.yml
@@ -4,7 +4,6 @@
   name: custom-frame-trait-overvolted-i-name
   description: custom-frame-trait-overvolted-i-description
   incompatible:
-    - OvervoltedII
     - UndervoltedI
     - UndervoltedII
   cost: -2
@@ -15,12 +14,58 @@
           drawRate: 0.75
 
 - type: backgroundTrait
+  id: HeavierFrame
+  name: custom-frame-trait-heavier-frame-name
+  description: custom-frame-trait-heavier-frame-description
+  cost: -4
+  effects:
+    - !type:OverrideCompsEffect
+      components:
+      - type: FootstepModifier
+        footstepSoundCollection:
+          path: /Audio/Mecha/sound_mecha_powerloader_step.ogg
+          params:
+            volume: -10
+      - type: HeavierFrameTrait
+        speedModifier: 0.95
+
+
+- type: backgroundTrait
+  id: CompromisedIntegrity
+  name: custom-frame-trait-compromised-integrity-name
+  description: custom-frame-trait-compromised-integrity-description
+  incompatible: 
+  - BetterInsulation
+  cost: -5
+  effects:
+    - !type:OverrideCompsEffect
+      components:
+      - type: SetDamageModifierTrait
+        damageModifierSetId: IPCWeaker
+
+- type: backgroundTrait
+  id: RadiationVulnerability
+  name: custom-frame-trait-radiation-vulnerability-name
+  description: custom-frame-trait-radiation-vulnerability-description
+  incompatible: 
+  - BetterInsulation
+  cost: -5
+  effects:
+    - !type:OverrideCompsEffect
+      components:
+      - type: SetDamageModifierTrait
+        damageModifierSetId: IPCRadiation
+      - type: GlitchOnIonStorm
+        duration: 60
+        ramp: 5
+
+- type: backgroundTrait
   id: CheapOilFilter
   name: custom-frame-trait-cheap-oil-filter-name
   description: custom-frame-trait-cheap-oil-filter-description
   incompatible: 
     - DeluxeOilFilter
-  cost: -3
+  cost: -6
   effects:
     - !type:OverrideCompsEffect
       components:
@@ -29,25 +74,52 @@
         bloodReductionAmount: 0
 
 - type: backgroundTrait
-  id: OvervoltedII
-  name: custom-frame-trait-overvolted-ii-name
-  description: custom-frame-trait-overvolted-ii-description
+  id: MicroreactorIncompatibility
+  name: custom-frame-trait-microreactor-incompatibility-name
+  description: custom-frame-trait-microreactor-incompatibility-description
+  cost: -8
+  incompatible: 
+  - BloodPowered
+  effects:
+    - !type:OverrideCompsEffect
+      components:
+      - type: MicroreactorIncompatibility
+
+#region Positive Traits
+
+- type: backgroundTrait
+  id: UndervoltedI
+  name: custom-frame-trait-undervolted-i-name
+  description: custom-frame-trait-undervolted-i-description
   incompatible:
-    - OvervoltedI
-    - UndervoltedI
     - UndervoltedII
-  cost: -4
+    - OvervoltedI
+  cost: 2
   effects:
     - !type:OverrideCompsEffect
       components:
         - type: PowerCellDraw
-          drawRate: 0.9
+          drawRate: 0.5
+
+- type: backgroundTrait
+  id: UndervoltedII
+  name: custom-frame-trait-undervolted-ii-name
+  description: custom-frame-trait-undervolted-ii-description
+  incompatible:
+    - UndervoltedI
+    - OvervoltedI
+  cost: 4
+  effects:
+    - !type:OverrideCompsEffect
+      components:
+        - type: PowerCellDraw
+          drawRate: 0.4
 
 - type: backgroundTrait
   id: BloodPowered
   name: custom-frame-trait-blood-powered-name
   description: custom-frame-trait-blood-powered-description
-  cost: -5
+  cost: 5
   incompatible: 
   - RepairNanites
   - DeluxeOilFilter
@@ -88,94 +160,6 @@
       - BatteryDrainer
 
 - type: backgroundTrait
-  id: CompromisedIntegrity
-  name: custom-frame-trait-compromised-integrity-name
-  description: custom-frame-trait-compromised-integrity-description
-  incompatible: 
-  - BetterInsulation
-  cost: -6
-  effects:
-    - !type:OverrideCompsEffect
-      components:
-      - type: SetDamageModifierTrait
-        damageModifierSetId: IPCWeaker
-
-- type: backgroundTrait
-  id: RadiationVulnerability
-  name: custom-frame-trait-radiation-vulnerability-name
-  description: custom-frame-trait-radiation-vulnerability-description
-  incompatible: 
-  - BetterInsulation
-  cost: -6
-  effects:
-    - !type:OverrideCompsEffect
-      components:
-      - type: SetDamageModifierTrait
-        damageModifierSetId: IPCRadiation
-      - type: GlitchOnIonStorm
-        duration: 60
-        ramp: 5
-
-- type: backgroundTrait
-  id: MicroreactorIncompatibility
-  name: custom-frame-trait-microreactor-incompatibility-name
-  description: custom-frame-trait-microreactor-incompatibility-description
-  cost: -6
-  incompatible: 
-  - BloodPowered
-  effects:
-    - !type:OverrideCompsEffect
-      components:
-      - type: MicroreactorIncompatibility
-
-#region Positive Traits
-
-- type: backgroundTrait
-  id: UndervoltedI
-  name: custom-frame-trait-undervolted-i-name
-  description: custom-frame-trait-undervolted-i-description
-  incompatible:
-    - UndervoltedII
-    - OvervoltedI
-    - OvervoltedII
-  cost: 2
-  effects:
-    - !type:OverrideCompsEffect
-      components:
-        - type: PowerCellDraw
-          drawRate: 0.5
-
-- type: backgroundTrait
-  id: UndervoltedII
-  name: custom-frame-trait-undervolted-ii-name
-  description: custom-frame-trait-undervolted-ii-description
-  incompatible:
-    - UndervoltedI
-    - OvervoltedI
-    - OvervoltedII
-  cost: 4
-  effects:
-    - !type:OverrideCompsEffect
-      components:
-        - type: PowerCellDraw
-          drawRate: 0.4
-
-- type: backgroundTrait
-  id: DeluxeOilFilter
-  name: custom-frame-trait-deluxe-oil-filter-name
-  description: custom-frame-trait-deluxe-oil-filter-description
-  incompatible: 
-    - CheapOilFilter
-    - BloodPowered
-  cost: 4
-  effects:
-    - !type:OverrideCompsEffect
-      components:
-      - type: ModifyBloodstreamTrait
-        bloodRefreshRate: 2.0
-        bloodReductionAmount: 0.33
-
-- type: backgroundTrait
   id: RepairNanites
   name: custom-frame-trait-repair-nanites-name
   description: custom-frame-trait-repair-nanites-description
@@ -194,12 +178,9 @@
         damageCap: 75
         damage:
           types:
-            Heat: -0.5
-            Shock: -0.5
-            Cold: -0.5
-            Blunt: -0.5
-            Piercing: -0.5
-            Slash: -0.5
+            Blunt: -1.0
+            Piercing: -1.0
+            Slash: -1.0
 
 - type: backgroundTrait
   id: BetterInsulation
@@ -214,8 +195,6 @@
       components:
       - type: Insulated
         coefficient: 0.4
-      - type: SetDamageModifierTrait
-        damageModifierSetId: IPCInsulated
 
 - type: backgroundTrait
   id: ModuleCompatibility
